- Added some missing WINELIB_NAME_AW definitions, types and messages - Small prototype fixes

+++ include/windef.h
@@ -450,10 +450,9 @@ typedef LRESULT CALLBACK (*WNDPROC)(HWND,UINT,WPARAM,LPARAM);
 /* The SIZE structure */
 typedef struct tagSIZE
 {
-    INT  cx;
-    INT  cy;
+    LONG cx;
+    LONG cy;
 } SIZE, *PSIZE, *LPSIZE;
-
 
 typedef SIZE SIZEL, *PSIZEL, *LPSIZEL;
